cleanup old test code
